cloud_service/requirements-dev.txt
# Dev / test dependencies — install with: pip install -r requirements-dev.txt
-r requirements.txt

pytest>=8.0
pytest-asyncio>=0.23
pytest-cov>=5.0
pytest-dotenv>=0.5        # loads .env.test via pyproject.toml env_files
httpx>=0.27               # async test client (also in requirements.txt)
aiosqlite>=0.20           # SQLite async driver for isolated test DB
python-dotenv>=1.0        # load .env.test in conftest fallback
